分离代码
